requirements.txt
APScheduler==3.6.3
apturl==0.5.2
asttokens==2.0.5
attrs==19.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.4
blis==0.2.4
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
ChatterBot==1.0.5
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cymem==2.0.6
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
executing==0.8.3
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.9.1
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
mathparse==0.1.2
matplotlib-inline==0.1.3
more-itertools==4.2.0
murmurhash==1.0.6
nest-asyncio==1.5.4
netifaces==0.10.4
nltk==3.7
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
Pint==0.18
pipreqs==0.4.11
plac==0.9.6
preshed==2.0.1
prompt-toolkit==3.0.28
protobuf==3.6.1
pure-eval==0.2.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-telegram-bot==13.11
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==1.0.5
stack-data==0.2.0
systemd-python==234
texttable==1.6.2
thinc==7.0.8
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wasabi==0.9.0
wcwidth==0.2.5
websocket-client==0.53.0
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
